fixed a lot of bugs, created a custom time picker to limit the choice of time, added the choice of the maximum date of delivery or doctor's appointment, the full functionality of the application was checked

+++ app/src/main/java/com/saimone/healthcareapp/BuyMedicineBookActivity.java
@@ -10,8 +10,6 @@ import android.widget.Button;
 import android.widget.EditText;
 import android.widget.Toast;
 
-import java.util.regex.Pattern;
-
 public class BuyMedicineBookActivity extends AppCompatActivity {
     EditText etFullName, etAddress, etPinCode, etContactno;
     Button bookButton;
@@ -28,18 +26,29 @@ public class BuyMedicineBookActivity extends AppCompatActivity {
         bookButton = findViewById(R.id.btnBMBBook);
 
         Intent it = getIntent();
-        String[] price = it.getStringExtra("price").split(Pattern.quote("$"));
+        String priceStr = it.getStringExtra("price");
+        int p1 = priceStr.indexOf(":") + 1;
+        int p2 = priceStr.indexOf("$", p1);
+        float price = Float.parseFloat(priceStr.substring(p1, p2));
         String date = it.getStringExtra("date");
 
         bookButton.setOnClickListener(view -> {
-            SharedPreferences sharedPreferences = getSharedPreferences("shared_prefs", Context.MODE_PRIVATE);
+            SharedPreferences sharedPreferences = getSharedPreferences("shared_preferences", Context.MODE_PRIVATE);
             String username = sharedPreferences.getString("username", "");
+            String fullname = etFullName.getText().toString();
+            String address = etAddress.getText().toString();
+            String pinCode = etPinCode.getText().toString();
+            String contactno = etContactno.getText().toString();
 
             try(Database db = new Database(getApplicationContext(), "healthcare", null, 1)) {
-                db.addOrder(username, etFullName.getText().toString(), etAddress.getText().toString(), Integer.parseInt(etPinCode.getText().toString()), etContactno.getText().toString(), date, "", Float.parseFloat(price[1]), "medicine");
-                db.removeCart(username, "medicine");
-                Toast.makeText(getApplicationContext(), "Your booking is done successfully", Toast.LENGTH_LONG).show();
-                startActivity(new Intent(BuyMedicineBookActivity.this, HomeActivity.class));
+                if (fullname.length() == 0 || address.length() == 0 || pinCode.length() == 0 || contactno.length() == 0) {
+                    Toast.makeText(getApplicationContext(), "Please enter all data", Toast.LENGTH_SHORT).show();
+                } else {
+                    db.addOrder(username, fullname, address, Integer.parseInt(pinCode), contactno, date, "", price, "medicine");
+                    db.removeCart(username, "medicine");
+                    Toast.makeText(getApplicationContext(), "Your booking is done successfully", Toast.LENGTH_LONG).show();
+                    startActivity(new Intent(BuyMedicineBookActivity.this, HomeActivity.class));
+                }
             } catch (Exception e) {
                 e.printStackTrace();
             }
